Move trainer.Feeder to lib.training.generator
