+++ lib/training/__init__.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import typing as T
 
 from .augmentation import ImageAugmentation
-from .generator import PreviewDataGenerator, TrainingDataGenerator
+from .generator import Feeder
 from .preview_cv import PreviewBuffer, TriggerType
 
 if T.TYPE_CHECKING:
